refactoring:
 - moved main_window class in a separated module
 - created directory various to hold icon src and other stuff
   that don't need to be on main
 - moved lang handling in a separated module
